Add the "bar-testchip" chip generator
This will add the various things to Rocket Chip that we're using for our
test chips.  Right now all it does is add another layer of heirarchy,
but it does pass the test cases and uses the pcad stuff that's going to
be necessary to make the new HTIF work.