Pass the path to ppkg-config to pconfigure runs
It turns out that Ben's failure could be reproduced on Travis, and was
just the result of me having ppkg-config in my PATH on my machines.  An
update to pconfigure allows the path to ppkg-config to be specified when
pconfigure is run (for situations like these).

I believe I don't need to set PKG_CONFIG_PATH here because it's already
exported, but we'll have to see!