Allow other core generators to use rocket-chip
I want to be able to have some core generators that are mostly based on
Rocket Chip, but do things like define a new top-level Chisel file.
There's two changes here:

 * CORE_TOP can be overridden.
 * CORE_ADDON_DIR can be set, which will be copied into Rocket Chip.

I've got a sample of this, but it's not working yet (presumably for
other reasons, this all seems pretty benign).