Add a crossbar "core generator"
The crossbar isn't really a core, but some external people want the
Nasti/AXI crossbar pulled out of Rocket Chip and instantiated as a
top-level design by itself.

This patch adds a new "core generator" which is really just a new top
level for Rocket Chip that does a few things differently.  Specifically,
it:

 * Adds some new Config entries (master and slave ports on the crossbar)
 * Adds a new top-level module (just a crossbar)
 * Defines no core-level test cases