Don't put CLOCK_PERIOD in the simulation scripts
I'm trying to avoid rocket-chip-isms in the rest of PLSI, as it's meant
to apply to other top-level designs.  Rather than defining CLOCK_PERIOD
as part of the simulation scripts, there's just a file in the Rocket
Chip addon that defines CLOCK_PEROID.

I'm going to have to figure out exactly what to do with clocking when
running the synthesis scripts, but for now I'll just live with this.