Ignore some VCS-specific Rocket Chip files in Verilator
OK, so I know I just said I didn't want rocket-chip-isms to slip into
the simulators but I'm just not sure how to get around this: VCS can't
parse the Verilator harness, and Verilator can't parse the VCS harness.

I think upstream is going to migrate to a single harness at some point,
but for now I'm just going to have to live with this.