Allow tarballs to exist on the filesystem, and do so for SAED32
The technology tarballs are big and it's a pain to have users constantly
fetch them.  This allows a Makefile.local or Makefile.project to specify
a variable that is the directory that contains technology tarballs.
PLSI will then go and extract from those tarballs rather than popping up
a "you should download these tarballs" message.