Bump Rocket Chip
They changed around a whole bunch of stuff upstream, but pretty much all
I needed to do here was to change some of the filenames that come out of
Rocket Chip's make invocation.

This was pretty much just a weekly update, but the SRAM stuff upstream
might work now so I can try it.