Allow users to pass a ".ucli" file to VCS
When we're running gate-level simulation on compilcated designs we need
to deal with un-initialized registers.  Since Chisel doesn't emit
X-tolerant Verilog the only way I can figure out how to initialize the
registers is to use Yunsup's "force" script, which is a Synopsys thing.

This commit just allows passing these sorts of arguments to VCS.  It's
handled with a magic, non-make-tracked file.  This is bad, but I'm just
going to have to live with it for now.