Read the TSMC documentation and produce a list of SRAMs
Last time I tried to do this I tried to read the TSMC documentation
myself and encode that information into a program that describes which
SRAMs are availiable.  This was a mess, because rules are confusing.

This time I decided to just teach my computer how to read the TSMC
documentation.  There's maybe 10k SRAMs availiable, but the new macro
compiler is fast enough that it doesn't really matter.  I manually
verified DefaultConfig and it maps well to the SRAMs, but I don't
actually have the rest of the memory compiler scripted yet so it's going
to take a bit.