Allow users of the Rocket Chip addon to set CORE_SIM_HARNESS
RocketChip has a Verilog top-level of the test harness that hard codes
ExampleHarness.  If users want to be able to use the PLSI Rocket Chip
addon system (which is essentially just Rocket Chip's own system), then
they'll need to change the name of ExampleHarness.  This system allows
users to provide a Makefile rule that generates their own test harness.