Add finer-grained makefile dependency variables
This makes it easier to rerun as few steps as needed.
If you have your yaml files split up by which action they affect
then you can set the new variables to point to the respective
yaml, and change something that only affects par will only rerun par,
or if you change something that affects par but run power it will be able
to rerun par, sim-par, and power.