Key history (#668)
Every run of the CLI now creates a {action}-output-history.yml file that contains a flattened version of the full output config, annotated with comments describing the files that modified said keys in the order they were provided as project files in the CLI call. This is done by creating a temporary file similar in fashion to the CLI driver tests.