Add support for process variation to MMMC Corners
This falls in a few nested categories.
ProcessVariation for front end of line (FEOL) variations of the devices
themselves we support indepdent nmos and pmos variation but only three
flavors slow, typical, and fast.
RCVariation for back end of line (BEOL) variations of the metal layers
we support a primary RC corner type and a coupling RC type which shows
up in multi-patterned processes. These support the standard 5 and 2
types respectively.